requirements.txt
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.4
Django==3.1.5
django-cors-headers==3.7.0
django-markdown==0.8.4
django-restql==0.11.0
djangorestframework==3.12.2
drf-spectacular==0.13.1
inflection==0.5.1
jsonschema==3.2.0
Markdown==3.3.3
mysqlclient==2.0.3
pycodestyle==2.6.0
pyPEG2==2.15.2
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
